GS-4456
Add support in the IntegratedProcessingUnitCotainer to deafult for pu.xml in the PDE project structure format. Makes life easier for eclipse users.

Now, META-INF/MANIFEST.MF/spring/pu.xml does not have to be in the class path, it can be in the root of the project directory.

Reviewer: Banon

svn path=/trunk/openspaces/; revision=24383


Former-commit-id: 115962351712f715c45d1cad46e61d059baff6e0