GS-4456
Add support in the IntegratedProcessingUnitCotainer to deafult for pu.xml in the PDE project structure format. Makes life easier for eclipse users.

Now, META-INF/MANIFEST.MF/spring/pu.xml does not have to be in the class path, it can be in the root of the project directory.

Reviewer: Banon

svn path=/trunk/openspaces/; revision=24389


Former-commit-id: 9af9dea835981e73dbc9061064fa7992a5f52a6e