GS-5982
When GS is installed under a path that contains white spaces, using the Maven that came with the GS installation fails.
The reason is that the bash command `dirname $0` breaks when using a pach with white spaces. To bypass this problem I use "`dirname \"$0\"`" which works in this situation.
In addition some changes to setenv.sh had to be done for it to handle paths with white spaces: surround ${JSHOMEDIR} with quotation marks (").

We need to test to see that the original functionality of setenv wasn't broken.

In addition, it seems that the unit scripts are not fully prepared to work with white space paths.

svn path=/trunk/gigaspaces/; revision=42916


Former-commit-id: 041c5aa9844b91a72b0a6f81048ef9a3de82a47d