GS-10441: Admin that recovers from network disconnection fails to fetch processing unit Space instance details.
DefaultSpace.addInstance schedules a ScheduledRuntimeFetcher to fetch SpaceInstance details (SpaceMode, etc.)

DefaultSpace.removeInstance cancels the future ScheduledRuntimeFetcher. But, it does not remove it from the map. When addInstance is called with the space instance that was discovered and disconnected, a new fetcher is started only if it is not in already in the map. Fix: remove it from the map so that next time it will be created.

svn path=/xap/trunk/openspaces/; revision=123557


Former-commit-id: b564831a02684bde491191d4b0e2eac870170db1