GS-10616
Test is taking wrong assumption, even though both partitions are embedded with the Jvm, from the router perspective, one of them is remote so the IncrementTask is serialized. Thus the test cannot assume one partition will get IncrementalTask with value 1 and the other with 2 which will generate average of 1.5

Changed IncrementalTask to behave according to this test assumption using static integer in it.

svn path=/xap/trunk/openspaces/; revision=130099


Former-commit-id: a1c863c91e3761c2569b2f1f3990b2de510006f2